bin/requirements.txt
pandas>=1.3.5
futures>=3.0
biopython>=1.81
tqdm>=4.66.1
